backend/requirements.txt
pandas
scikit-learn
numpy
python-dotenv
requests
flask
flask-cors